start list testing
