+++ src/main/java/frc/robot/RobotContainer.java
@@ -33,7 +33,6 @@ import frc.robot.commands.BabyBirdCommand;
 import frc.robot.commands.DefaultElevatorCommand;
 import frc.robot.commands.DefaultTurretCommand;
 import frc.robot.commands.EjectIntakeCommand;
-import frc.robot.commands.DrivetrainTestCommand;
 import frc.robot.commands.FieldOrientedDriveCommand;
 import frc.robot.commands.ManualShootCommand;
 import frc.robot.commands.SpeakerOrBlinkCommand;
@@ -41,7 +40,6 @@ import frc.robot.commands.TuneSpeakerCommand;
 import frc.robot.commands.led.DefaultLEDCommand;
 import frc.robot.commands.led.LEDBlinkCommand;
 import frc.robot.commands.led.LEDBootAnimationCommand;
-import frc.robot.commands.led.LEDProgressBarCommand;
 import frc.robot.controls.ControlBindings;
 import frc.robot.controls.JoystickControlBindings;
 import frc.robot.controls.XBoxControlBindings;
@@ -279,11 +277,6 @@ public class RobotContainer {
     tab.add("Tur Roll Dynam Fwd", turretSubsystem.sysIdRollerDynamicCommand(kForward)).withPosition(columnIndex, 2);
     tab.add("Tur Roll Dynam Rev", turretSubsystem.sysIdRollerDynamicCommand(kReverse)).withPosition(columnIndex, 3);
   
-    // Testing
-    tab = Shuffleboard.getTab("Testing");
-
-    tab.add("Drivetrain test", new DrivetrainTestCommand(drivetrain)).withPosition(0, 0);
-    tab.add("Complete Test", new LEDProgressBarCommand(ledSubsystem)).withPosition(0, 1);
   }
 
   public void setAlliance(Alliance alliance) {
