--- src/main/java/frc/robot/commands/DrivetrainTestCommand.java
@@ -1,75 +1,0 @@
-package frc.robot.commands;
-
-import static edu.wpi.first.units.Units.MetersPerSecond;
-import static edu.wpi.first.units.Units.MetersPerSecondPerSecond;
-import static edu.wpi.first.units.Units.RadiansPerSecond;
-import static edu.wpi.first.units.Units.Second;
-import static frc.robot.Constants.DrivetrainConstants.MAX_ANGULAR_VELOCITY;
-import static frc.robot.Constants.DrivetrainConstants.MAX_VELOCITY;
-import static frc.robot.Constants.TeleopDriveConstants.ROTATIONAL_DEADBAND;
-import static frc.robot.Constants.TeleopDriveConstants.ROTATION_RATE_LIMIT;
-import static frc.robot.Constants.TeleopDriveConstants.TRANSLATION_DEADBAND;
-import static frc.robot.Constants.TeleopDriveConstants.TRANSLATION_RATE_LIMIT;
-
-import com.ctre.phoenix6.mechanisms.swerve.SwerveModule.DriveRequestType;
-import com.ctre.phoenix6.mechanisms.swerve.SwerveModule.SteerRequestType;
-import com.ctre.phoenix6.mechanisms.swerve.SwerveRequest;
-import com.pathplanner.lib.util.ChassisSpeedsRateLimiter;
-
-import edu.wpi.first.math.kinematics.ChassisSpeeds;
-import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.CommandSwerveDrivetrain;
-
-/**
- * Command for making sure the drivetrain is working properly
- */
-public class DrivetrainTestCommand extends Command {
-  private final CommandSwerveDrivetrain drivetrainSubsystem;
-  
-  private final ChassisSpeedsRateLimiter rateLimiter = new ChassisSpeedsRateLimiter(
-      TRANSLATION_RATE_LIMIT.in(MetersPerSecondPerSecond), ROTATION_RATE_LIMIT.in(RadiansPerSecond.per(Second)));
-
-      private final SwerveRequest.FieldCentric drive = new SwerveRequest.FieldCentric()
-      .withDeadband(MAX_VELOCITY.times(TRANSLATION_DEADBAND).in(MetersPerSecond))
-      .withRotationalDeadband(MAX_ANGULAR_VELOCITY.times(ROTATIONAL_DEADBAND).in(RadiansPerSecond))
-      .withDriveRequestType(DriveRequestType.Velocity)
-      .withSteerRequestType(SteerRequestType.MotionMagic);
-      
-  // Desired chassis speeds. Defined here to prevent reallocation.
-  private ChassisSpeeds desiredChassisSpeeds = new ChassisSpeeds();
-
-  /**
-   * Constructor
-   * @param drivetrainSubsystem drivetrain
-   */
-  public DrivetrainTestCommand(
-      CommandSwerveDrivetrain drivetrainSubsystem) {
-    this.drivetrainSubsystem = drivetrainSubsystem;
-
-    addRequirements(drivetrainSubsystem);
-  }
-
-  @Override
-  public void initialize() {
-    // Reset the slew rate limiters, in case the robot is already moving
-    rateLimiter.reset(drivetrainSubsystem.getCurrentFieldChassisSpeeds());
-  }
-
-  @Override
-  public void execute() {
-    desiredChassisSpeeds.vxMetersPerSecond = 0.1;
-    desiredChassisSpeeds.vyMetersPerSecond = 0.1;
-    desiredChassisSpeeds.omegaRadiansPerSecond = 0.1;
-    var limitedCassisSpeeds = rateLimiter.calculate(desiredChassisSpeeds);
-    drivetrainSubsystem.setControl(drive
-        .withVelocityX(limitedCassisSpeeds.vxMetersPerSecond)
-        .withVelocityY(limitedCassisSpeeds.vyMetersPerSecond)
-        .withRotationalRate(limitedCassisSpeeds.omegaRadiansPerSecond));
-  }
-
-  @Override
-  public void end(boolean interrupted) {
-    drivetrainSubsystem.setControl(new SwerveRequest.ApplyChassisSpeeds().withSpeeds(new ChassisSpeeds()));
-  }
-
-}
